Split up test_tuning.py
